requirements.txt
flask==3.0
flask_restful==0.3.10
flask_sqlalchemy==3.1.1
flask_migrate==4.0.5
flask-Mail==2.1.5
flask_marshmallow==1.2
flask-WTF==0.14.2
flask-login==0.6.3
email_validator==2.1.0
marshmallow-sqlalchemy==1.0.0
python-slugify==3.0.2
SQLAlchemy==2.0.25
text-unidecode==1.2
WTForms==3.1.2
python-dateutil==2.8.1
python-editor==1.0.4
PyPDF2==3.0.1
pytesseract==0.3.10
tesseract-ocr==0.0.1
python-docx==1.1.0